src/db_orchestration/requirements.txt
# Dépendances Orchestrateur (Runner)
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg[binary]
pysqlite3-binary
requests
httpx
aiohttp
tenacity
python-dateutil
pytz
python-json-logger
prometheus-client
ollama
bleach
markdownify
Jinja2
PyPDF2
latexcodec
pylatexenc
python-docx
markdown
chevron
anthropic
spacy
keybert
sentence-transformers
langdetect
streamlit
streamlit-antd-components
